feat: update visualizer for better display
We add _create_axes_simple() and _create_axes_fancy(), to offer two
different visualization options. The latter is the default, as it is the
'prettier' of the two, but we can fallback if we find bugs with
my (crude) logic in it.

We also add a bit of pydoc comments.