refactor: significant changes to visualizer
Various changes to try and make the visualizer class a bit cleaner.
There is still quite a bit that could be done :(.

- Renaming ExperimentVisualizer to SpectraStackVisualizer, to try and
clarify what it is doing.
- image_dims -> image_axes, to be more explicit that it is the axis ids.
- created method _visualize_spectral_topography(), which is a generic
version of set_image. set_image is now set_selection_image(), and it
calls the generic method and adds a rectangle.
- verify_dataset() is now _verify_and_setup_dataset().
- _visualize_spectra() is now  _visualize_spectral_data(), to be
more generic (we could ovveride this method to visualize in a different way).
- _scale_bar() was turned into a method create_scale_bar(), outside of
the class, that is more generic.
- added method create_color_bar(), to also generically...add a color bar.